chat_backend/production-requirements.txt
flask
dotenv
google-generativeai
waitress
langchain
langchain-google-genai
langchain-community
flask-cors
python-dotenv
faiss-cpu
